Release package for Gluster 3.9
This package is not supposed to get installed when a users does

    # yum install centos-release-gluster

Instead, a user needs to explicitly install centos-release-gluster39.

The Gluster 3.9 version is a Short-Term-Maintenance version and will
become obsolete when Gluster 3.10 is available. We do not want to burden
users too much, and only have the Long-Term-Maintenance releases
installed by default.

Applications (like oVirt) that depend on centos-release-gluster with a
recent version can still try out running with Gluster 3.9. In these
cases both centos-release-gluster and centos-release-gluster39 need to
be installed. This will provide glusterfs-3.9.x and additional
dependencies as required.

Signed-off-by: Niels de Vos <ndevos@redhat.com>